Check what OS is used before unlink sqlite database

+++ app/Traits/InitDatabaseTrait.php
@@ -18,17 +18,34 @@ trait InitDatabaseTrait
         }
 
         $db = $this->app->make('db')->connection();
-        if(!file_exists($db->getDatabaseName())) {
+        if (!file_exists($db->getDatabaseName())) {
             touch($db->getDatabaseName());
         }
 
         copy($db->getDatabaseName(), $db->getDatabaseName() . static::$backupExtension);
-        unlink($db->getDatabaseName());
+
+        $this->unlinkDatabase($db);
+
         touch($db->getDatabaseName());
 
         $this->artisan('migrate');
         $this->seed(DatabaseSeeder::class);
+
         $this->beforeApplicationDestroyed([$this, 'restoreDatabase']);
+    }
+
+    public function unlinkDatabase($db)
+    {
+        if (file_exists($db->getDatabaseName())) {
+            if (strtoupper(substr(PHP_OS, 0, 3)) === 'WIN') {
+                var_dump( $db->getDatabaseName() );
+                chmod($db->getDatabaseName(), 0777);
+                unlink(basename($db->getDatabaseName()));
+            } else {
+                chmod($db->getDatabaseName(), 0777);
+                unlink($db->getDatabaseName());
+            }
+        }
     }
 
     /**
@@ -41,6 +58,7 @@ trait InitDatabaseTrait
             $this->refreshApplication();
         }
         $db = $this->app->make('db')->connection();
+
         copy($db->getDatabaseName() . static::$backupExtension,
             $db->getDatabaseName());
         unlink($db->getDatabaseName() . static::$backupExtension);
